feat: ✨ Add Native Paper and navigation
Navigation is used for the sample app to add multiple
screens showing usage of UI components. An example
has been done but should be removed as soon as we
get real components.

Note that for now, we got an error when trying to use
the wrapped Button component from our lib, couldn't
figure out why yet.